+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 
 #### Fixes
 * fixed disappearing values on some wizard params
+* fixed complex type wizard param's nesting